# openapi.yaml
openapi: 3.0.0
paths:
...
  /residents/{residentUid}/beverage-records:
    get:
      operationId: getBeverageRecords
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      responses:
        "200":
          description: 飲料記録一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeverageRecordListResponseDto"
      security:
        - JWT-auth: []
      summary: 利用者の飲料記録一覧を取得
      tags: &a9
        - beverage-records
    post:
      operationId: createBeverageRecord
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeverageRecordCreateInputDto"
      responses:
        "201":
          description: 飲料記録の作成に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeverageRecordDto"
      security:
        - JWT-auth: []
      summary: 飲料記録を作成
      tags: *a9
  /residents/{residentUid}/beverage-records/{uid}:
    patch:
      operationId: updateBeverageRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeverageRecordUpdateInputDto"
      responses:
        "200":
          description: 飲料記録の更新に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeverageRecordDto"
      security:
        - JWT-auth: []
      summary: 飲料記録を更新
      tags: *a9
    delete:
      operationId: deleteBeverageRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 飲料記録の削除に成功
      security:
        - JWT-auth: []
      summary: 飲料記録を削除
      tags: *a9
  /residents/{residentUid}/beverage-records/{uid}/transcription:
    get:
      operationId: getBeverageRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 飲料記録の文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 飲料記録の文字起こしを取得
      tags: *a9
    patch:
      operationId: appendBeverageRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 飲料記録の文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 飲料記録の文字起こしを追記
      tags: *a9
    put:
      operationId: updateBeverageRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 飲料記録の文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 飲料記録の文字起こしを置換
      tags: *a9
    delete:
      operationId: deleteBeverageRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 飲料記録の文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: 飲料記録の文字起こしを削除
      tags: *a9
  /residents/{residentUid}/beverage-records/{uid}/extract:
    post:
      operationId: extractBeverageRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 飲料記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 飲料記録からの情報抽出に成功
          content:
            application/json:
              schema:
                type: string
      security:
        - JWT-auth: []
      summary: 飲料記録から情報を抽出
      tags: *a9
...
  /assessments/{uid}/transcription:
    get:
      operationId: getTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを取得
      tags: *a10
    patch:
      operationId: appendTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを追記
      tags: *a10
    put:
      operationId: updateTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを置換
      tags: *a10
    delete:
      operationId: deleteTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを削除
      tags: *a10
...
info:
  title: Nagara Care API
  description: ""
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    TranscriptionDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
    TranscriptionInputDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
...
    BeverageRecordDto:
      type: object
      properties:
        uid:
          type: string
          description: UID
        tenantUid:
          type: string
          description: テナントUID
        caregiverUid:
          type: string
          description: 介護者UID
        residentUid:
          type: string
          description: 利用者UID
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
        notes:
          type: string
          description: メモ
        beverageType:
          type: string
          description: 飲み物の種類
          enum:
            - WATER
            - TEA
            - OTHER
        volume:
          type: number
          description: 飲み物の量（ml）
      required:
        - uid
        - tenantUid
        - caregiverUid
        - residentUid
        - recordedAt
        - notes
        - beverageType
        - volume
    BeverageRecordListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BeverageRecordDto"
        total:
          type: number
          description: 総件数
      required:
        - items
        - total
    BeverageRecordCreateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        beverageType:
          type: string
          description: 飲み物の種類
          enum:
            - WATER
            - TEA
            - OTHER
          example: WATER
        volume:
          type: number
          description: 飲み物の量（ml）
          minimum: 0
          example: 200
      required:
        - recordedAt
        - beverageType
        - volume
    BeverageRecordUpdateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        beverageType:
          type: string
          description: 飲み物の種類
          enum:
            - WATER
            - TEA
            - OTHER
          example: WATER
        volume:
          type: number
          description: 飲み物の量（ml）
          minimum: 0
          example: 200
      required:
        - recordedAt
        - beverageType
        - volume
...